app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.11.5
langchain==0.3.25
langchain-community==0.3.24
langchain-core
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
python-dotenv==1.0.0
python-multipart==0.0.6
pytest==8.4.1
pytest-asyncio==0.23.6
websockets==15.0.1